+++ include/ntop_typedefs.h
@@ -349,7 +349,7 @@ typedef struct zmq_remote_stats {
   char remote_probe_version[64], remote_probe_os[64];
   char remote_probe_license[64], remote_probe_edition[64];
   char remote_probe_maintenance[64];
-  u_int8_t source_id, num_exporters;
+  u_int32_t source_id, num_exporters;
   u_int64_t remote_bytes, remote_pkts, num_flow_exports;
   u_int32_t remote_ifspeed, remote_time, local_time, avg_bps, avg_pps;
   u_int32_t remote_lifetime_timeout, remote_idle_timeout,
